==> notebooks/9.2-upload_diario.ipynb <==
## set covid19_datasciente as path for python find bulletin package
import sys
from pathlib import Path
from os import getcwd, remove, chdir, rename
from os.path import join, basename, isdir, isfile

sys.path.append(str(Path(getcwd()).parent))
current_dir = getcwd()
print(current_dir)

import pandas as pd

from datetime import datetime, date, timedelta
from time import sleep

hoje = pd.to_datetime(date.today())
ontem = hoje - timedelta(1)
anteontem = ontem - timedelta(1)

basepath = join('C:\\', 'SESA')
path = join('C:\\', 'SESA', 'covid19datascience-dev', 'scripts_python', 'upload_diario.py')

if ( (hoje.day_name() == 'Saturday') or (hoje.day_name() == 'Sunday') ):
    print('Executando o UPLOAD (FDS)....')
    # upload files
    exec(open(path).read())
else:
    now = datetime.now()
    todayXXpm = now.replace(hour=14, minute=30, second=0, microsecond=0)

    finished = False
    while not finished:
        now = datetime.now()

        if ( isfile(join(basepath, 'upload=true.txt')) or (now > todayXXpm) ):
            #upload files
            print('Executando o UPLOAD....')
            exec(open(path).read())
            
            finished = True
            rename( join(basepath, 'upload=true.txt'), join(basepath, 'upload=finished.txt') )
        else:
            print('Waiting...')
            sleep(120)

exit
